// src/app/_core/models/customer-park-item-vo.ts
import { ReferenceDataVO } from './reference-data-vo';
import { CustomerParkHardwareAssociatedLineVO } from './customer-park-hardware-associated-line-vo';
import { CustomerParkItemServicesOptionsVO, CustomerParkItemTerminalMobileVO,
         CustomerParkItemContractVO } from './models';

export interface CustomerParkItemVO {
  id: number;
  name: string;
  creatorId: number;
  webServiceIdentifier: string;
  customerId: number;
  customerBeneficiairyId: number;
  modifierId: number;
  webService: number;
  customerIdentifier: string;
  customerLastNameFirstName: string;
  beneficiaryFirstName: string;
  beneficiaryLastName: string;
  account: string;
  comment: string;
  modificationDate: Date;
  universe: string;
  userId: number;
  userName: string;
  num50d: string;
  dateSubscription: Date;
  dateCancellation: Date;
  dateUnnumbered: Date;
  customerParkItemServicesOptions: CustomerParkItemServicesOptionsVO[];
  customerParkItemTerminalMobile: CustomerParkItemTerminalMobileVO[];
  customerParkItemContract: CustomerParkItemContractVO;
  numDoss: string;
  libelleContract: string;
  codeContract: string;
  numClient: string;
  lastUpdatedDate: Date;
  reason: string;
  nicheContractStatus: string;
  newDateSubscription: Date;
  benefStatus: number;
  horsForfait: string;
  rangRenewal: string;
  messageExcept: string;
  offers: [];
  nameCustomerFromWebservice: string;
  civilityFromWebservice: string;
  address1FromWebservice: string;
  address2FromWebservice: string;
  address3FromWebservice: string;
  address4FromWebservice: string;
  countryFromWebservice: string;
  cityFromWebservice: string;
  postalCodeFromWebservice: string;
  modePaymentFromWebservice: string;
  cleRIBFromWebservice: string;
  codeGuichetFromWebservice: string;
  codeBanqueFromWebservice: string;
  nomTitulaireFromWebservice: string;
  numCompteFromWebservice: string;
  newNicheContractStatus: string;
  isAccountFoundInPeniche: boolean;
  newNumber: string;
  lineOrigin: ReferenceDataVO;
  authorizedRenewal: boolean;
  hasRenewalPeriod: boolean;
  isRestrictedOffer: boolean;
  renewalLabel: string;
  dateRenewal: Date;
  renewalMobileClassification: number;
  lineHolder: ReferenceDataVO;
  isParnasseNumber: boolean;
  linkedParkItemId: number;
  linkedParkItemLabel: string;
  lineHolderName: string;	
  renewalBeginDate: Date;
  renewalEndDate: Date;
  commentRenewal: string;
  customerParkHardwareAssociatedLineVO : CustomerParkHardwareAssociatedLineVO;
}